feature(transform) Isolate the transformation engine, make it extensible
